Struct aws_sdk_memorydb::client::fluent_builders::CreateCluster

pub struct CreateCluster<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateCluster<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateClusterOutput, SdkError<CreateClusterError>> where
R::Policy: SmithyRetryPolicy<CreateClusterInputOperationOutputAlias, CreateClusterOutput, CreateClusterError, CreateClusterInputOperationRetryAlias>,

pub fn cluster_name(self, inp: impl Into<String>) -> Self

The name of the cluster. This value must be unique as it also serves as the cluster identifier.

pub fn set_cluster_name(self, input: Option<String>) -> Self

pub fn node_type(self, inp: impl Into<String>) -> Self

The compute and memory capacity of the nodes in the cluster.

pub fn set_node_type(self, input: Option<String>) -> Self

pub fn parameter_group_name(self, inp: impl Into<String>) -> Self

The name of the parameter group associated with the cluster.

pub fn set_parameter_group_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

An optional description of the cluster.

pub fn set_description(self, input: Option<String>) -> Self

pub fn num_shards(self, inp: i32) -> Self

The number of shards the cluster will contain. The default value is 1.

pub fn set_num_shards(self, input: Option<i32>) -> Self

pub fn num_replicas_per_shard(self, inp: i32) -> Self

The number of replicas to apply to each shard. The default value is 1. The maximum is 5.

pub fn set_num_replicas_per_shard(self, input: Option<i32>) -> Self

pub fn subnet_group_name(self, inp: impl Into<String>) -> Self

The name of the subnet group to be used for the cluster.

pub fn set_subnet_group_name(self, input: Option<String>) -> Self

pub fn security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to SecurityGroupIds.

To override the contents of this collection use set_security_group_ids.

A list of security group names to associate with this cluster.

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn maintenance_window(self, inp: impl Into<String>) -> Self

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

pub fn set_maintenance_window(self, input: Option<String>) -> Self

pub fn port(self, inp: i32) -> Self

The port number on which each of the nodes accepts connections.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn sns_topic_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

pub fn set_sns_topic_arn(self, input: Option<String>) -> Self

pub fn tls_enabled(self, inp: bool) -> Self

A flag to enable in-transit encryption on the cluster.

pub fn set_tls_enabled(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

The ID of the KMS key used to encrypt the cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn snapshot_arns(self, inp: impl Into<String>) -> Self

Appends an item to SnapshotArns.

To override the contents of this collection use set_snapshot_arns.

A list of Amazon Resource Names (ARN) that uniquely identify the RDB snapshot files stored in Amazon S3. The snapshot files are used to populate the new cluster. The Amazon S3 object name in the ARN cannot contain any commas.

pub fn set_snapshot_arns(self, input: Option<Vec<String>>) -> Self

pub fn snapshot_name(self, inp: impl Into<String>) -> Self

The name of a snapshot from which to restore data into the new cluster. The snapshot status changes to restoring while the new cluster is being created.

pub fn set_snapshot_name(self, input: Option<String>) -> Self

pub fn snapshot_retention_limit(self, inp: i32) -> Self

The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

pub fn set_snapshot_retention_limit(self, input: Option<i32>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A list of tags to be added to this resource. Tags are comma-separated key,value pairs (e.g. Key=myKey, Value=myKeyValue. You can include multiple tags as shown following: Key=myKey, Value=myKeyValue Key=mySecondKey, Value=mySecondKeyValue.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn snapshot_window(self, inp: impl Into<String>) -> Self

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard.

Example: 05:00-09:00

If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

pub fn set_snapshot_window(self, input: Option<String>) -> Self

pub fn acl_name(self, inp: impl Into<String>) -> Self

The name of the Access Control List to associate with the cluster.

pub fn set_acl_name(self, input: Option<String>) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

The version number of the Redis engine to be used for the cluster.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, inp: bool) -> Self

When set to true, the cluster will automatically receive minor engine version upgrades after launch.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self