Module aws_sdk_memorydb::model

Modules

ac_ls_update_status

See AcLsUpdateStatus

acl

See Acl

acl_pending_changes

See AclPendingChanges

authentication

See Authentication

authentication_mode

See AuthenticationMode

availability_zone

See AvailabilityZone

cluster

See Cluster

cluster_configuration

See ClusterConfiguration

cluster_pending_updates

See ClusterPendingUpdates

endpoint

See Endpoint

engine_version_info

See EngineVersionInfo

event

See Event

filter

See Filter

node

See Node

parameter

See Parameter

parameter_group

See ParameterGroup

parameter_name_value

See ParameterNameValue

pending_modified_service_update

See PendingModifiedServiceUpdate

replica_configuration_request

See ReplicaConfigurationRequest

resharding_status

See ReshardingStatus

security_group_membership

See SecurityGroupMembership

service_update

See ServiceUpdate

service_update_request

See ServiceUpdateRequest

shard

See Shard

shard_configuration

See ShardConfiguration

shard_configuration_request

See ShardConfigurationRequest

shard_detail

See ShardDetail

slot_migration

See SlotMigration

snapshot

See Snapshot

subnet

See Subnet

subnet_group

See SubnetGroup

tag

See Tag

unprocessed_cluster

See UnprocessedCluster

user

See User

Structs

AcLsUpdateStatus

The status of the ACL update

Acl

An Access Control List. You can authenticate users with Access Contol Lists.

AclPendingChanges

Returns the updates being applied to the ACL.

Authentication

Denotes the user's authentication properties, such as whether it requires a password to authenticate. Used in output responses.

AuthenticationMode

Denotes the user's authentication properties, such as whether it requires a password to authenticate. Used in output responses.

AvailabilityZone

Indicates if the cluster has a Multi-AZ configuration (multiaz) or not (singleaz).

Cluster

Contains all of the attributes of a specific cluster.

ClusterConfiguration

A list of cluster configuration options.

ClusterPendingUpdates

A list of updates being applied to the cluster

Endpoint

Represents the information required for client programs to connect to the cluster and its nodes.

EngineVersionInfo

Provides details of the Redis engine version

Event

Represents a single occurrence of something interesting within the system. Some examples of events are creating a cluster or adding or removing a

Filter

Used to streamline results of a search based on the property being filtered.

Node

Represents an individual node within a cluster. Each node runs its own instance of the cluster's protocol-compliant caching software.

Parameter

Describes an individual setting that controls some aspect of MemoryDB behavior.

ParameterGroup

Represents the output of a CreateParameterGroup operation. A parameter group represents a combination of specific values for the parameters that are passed to the engine software during startup.

ParameterNameValue

Describes a name-value pair that is used to update the value of a parameter.

PendingModifiedServiceUpdate

Update action that has yet to be processed for the corresponding apply/stop request

ReplicaConfigurationRequest

A request to configure the number of replicas in a shard

ReshardingStatus

The status of the online resharding

SecurityGroupMembership

Represents a single security group and its status.

ServiceUpdate

An update that you can apply to your MemoryDB clusters.

ServiceUpdateRequest

A request to apply a service update

Shard

Represents a collection of nodes in a cluster. One node in the node group is the read/write primary node. All the other nodes are read-only Replica nodes.

ShardConfiguration

Shard configuration options. Each shard configuration has the following: Slots and ReplicaCount.

ShardConfigurationRequest

A request to configure the sharding properties of a cluster

ShardDetail

Provides details of a shard in a snapshot

SlotMigration

Represents the progress of an online resharding operation.

Snapshot

Represents a copy of an entire cluster as of the time when the snapshot was taken.

Subnet

Represents the subnet associated with a cluster. This parameter refers to subnets defined in Amazon Virtual Private Cloud (Amazon VPC) and used with MemoryDB.

SubnetGroup

Represents the output of one of the following operations:

Tag

A tag that can be added to an MemoryDB resource. Tags are composed of a Key/Value pair. You can use tags to categorize and track all your MemoryDB resources.

UnprocessedCluster

A cluster whose updates have failed

User

You create users and assign them specific permissions by using an access string. You assign the users to Access Control Lists aligned with a specific role (administrators, human resources) that are then deployed to one or more MemoryDB clusters.

Enums

AuthenticationType

AzStatus

InputAuthenticationType

ServiceUpdateStatus

ServiceUpdateType

SourceType