Module aws_sdk_memorydb::operation

Structs

BatchUpdateCluster

Apply the service update to a list of clusters supplied. For more information on service updates and applying them, see Applying the service updates.

CopySnapshot

Makes a copy of an existing snapshot.

CreateACL

Creates an Access Control List. For more information, see Authenticating users with Access Contol Lists (ACLs).

CreateCluster

Creates a cluster. All nodes in the cluster run the same protocol-compliant engine software.

CreateParameterGroup

Creates a new MemoryDB parameter group. A parameter group is a collection of parameters and their values that are applied to all of the nodes in any cluster. For

CreateSnapshot

Creates a copy of an entire cluster at a specific moment in time.

CreateSubnetGroup

Creates a subnet group. A subnet group is a collection of subnets (typically private) that you can designate for your clusters running in an Amazon Virtual Private Cloud (VPC) environment.

CreateUser

Creates a MemoryDB user. For more information, see Authenticating users with Access Contol Lists (ACLs).

DeleteACL

Deletes an Access Control List. The ACL must first be disassociated from the cluster before it can be deleted. For more information, see Authenticating users with Access Contol Lists (ACLs).

DeleteCluster

Deletes a cluster. It also deletes all associated nodes and node endpoints

DeleteParameterGroup

Deletes the specified parameter group. You cannot delete a parameter group if it is associated with any clusters.

DeleteSnapshot

Deletes an existing snapshot. When you receive a successful response from this operation, MemoryDB immediately begins deleting the snapshot; you cannot cancel or revert this operation.

DeleteSubnetGroup

Deletes a subnet group. You cannot delete a default subnet group or one that is associated with any clusters.

DeleteUser

Deletes a user. The user will be removed from all ACLs and in turn removed from all clusters.

DescribeACLs

Returns a list of ACLs

DescribeClusters

Returns information about all provisioned clusters if no cluster identifier is specified, or about a specific cluster if a cluster name is supplied.

DescribeEngineVersions

Returns a list of the available Redis engine versions.

DescribeEvents

Returns events related to clusters, security groups, and parameter groups. You can obtain events specific to a particular cluster, security group, or parameter group by providing the name as a parameter.

DescribeParameterGroups

Returns a list of parameter group descriptions. If a parameter group name is specified, the list contains only the descriptions for that group.

DescribeParameters

Returns the detailed parameter list for a particular parameter group.

DescribeServiceUpdates

Returns details of the service updates

DescribeSnapshots

Returns information about cluster snapshots. By default, DescribeSnapshots lists all of your snapshots; it can optionally describe a single snapshot,

DescribeSubnetGroups

Returns a list of subnet group descriptions. If a subnet group name is specified, the list contains only the description of that group.

DescribeUsers

Returns a list of users.

FailoverShard

Used to failover a shard

ListAllowedNodeTypeUpdates

Lists all available node types that you can scale to from your cluster's current node type.

ListTags

Lists all tags currently on a named resource.

ResetParameterGroup

Modifies the parameters of a parameter group to the engine or system default value. You can reset specific parameters by submitting a list of parameter names. To reset the entire parameter group, specify the AllParameters and ParameterGroupName parameters.

TagResource

A tag is a key-value pair where the key and value are case-sensitive. You can use tags to categorize and track all your MemoryDB resources.

UntagResource

Use this operation to remove tags on a resource

UpdateACL

Changes the list of users that belong to the Access Control List.

UpdateCluster

Modifies the settings for a cluster. You can use this operation to change one or more cluster configuration settings by specifying the settings and the new values.

UpdateParameterGroup

Updates the parameters of a parameter group. You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs.

UpdateSubnetGroup

Updates a subnet group. For more information, see Updating a subnet group

UpdateUser

Changes user password(s) and/or access string.